# src/main/resources/static/openapi/skillroute-api.yaml
openapi: 3.0.3
info:
  title: SkillRoute REST API
  version: "1.0"
  description: "REST API для AJAX-запросов SkillRoute: регистрация, пароль, чат, поиск навыков и синхронизация с GitHub."
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки
...
paths:
...
  /chat/{id}/messages:
    get:
      tags:
        - Чат
      summary: Получить сообщения чата
      description: Возвращает данные диалога и список сообщений.
      operationId: getChatMessages
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор чата
          schema:
            type: integer
            format: int64
          example: 10
      responses:
        '200':
          description: Диалог найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                chat:
                  summary: Диалог по отклику
                  value:
                    chatId: 10
                    opponentName: Арина
                    companyView: true
                    studentId: 7
                    vacancyId: 4
                    vacancyName: Java-разработчик
                    studentVacancyStatus: INTERVIEW
                    closed: false
                    messages:
                      - id: 1
                        senderId: 3
                        senderName: SkillRoute Company
                        text: Здравствуйте, приглашаем вас на интервью.
                        createdAt: 18.05.2026 14:20
                        mine: true
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /chat/{id}/send:
    post:
      tags:
        - Чат
      summary: Отправить сообщение в чат
      description: Добавляет новое сообщение в диалог и возвращает созданное сообщение.
      operationId: sendChatMessage
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор чата
          schema:
            type: integer
            format: int64
          example: 10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageRequest'
            examples:
              message:
                summary: Новое сообщение
                value:
                  text: Спасибо, мне удобно завтра после 15:00.
      responses:
        '200':
          description: Сообщение отправлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              examples:
                sent:
                  summary: Созданное сообщение
                  value:
                    id: 15
                    senderId: 7
                    senderName: Арина
                    text: Спасибо, мне удобно завтра после 15:00.
                    createdAt: 18.05.2026 15:10
                    mine: true
        '400':
          $ref: '#/components/responses/MessageValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  schemas:
...
    MessageRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 500
          description: Текст сообщения
          example: Здравствуйте, мне интересно ваше предложение.
          x-field-extra-annotation: '@jakarta.validation.constraints.NotBlank(message = "Текст сообщения не может быть пустым")'
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Общее сообщение об ошибке
          example: Проверьте корректность заполнения формы
        errorCode:
          type: string
          description: Код ошибки для фронтенда
          example: VALIDATION_ERROR
        fields:
          type: object
          description: Ошибки по конкретным полям формы
          additionalProperties:
            type: string
          example:
            email: Email обязателен
            password: Пароль должен содержать минимум 8 символов
...
    ChatResponse:
      type: object
      properties:
        chatId:
          type: integer
          format: int64
          description: Идентификатор чата
          example: 10
        opponentName:
          type: string
          description: Имя собеседника
          example: Арина
        messages:
          type: array
          description: Сообщения диалога
          items:
            $ref: '#/components/schemas/MessageResponse'
        companyView:
          type: boolean
          description: Открыт ли чат со стороны компании
          example: true
        studentId:
          type: integer
          format: int64
          description: Идентификатор студента в отклике
          example: 7
        vacancyId:
          type: integer
          format: int64
          description: Идентификатор вакансии
          example: 4
        vacancyName:
          type: string
          description: Название вакансии
          example: Java-разработчик
        studentVacancyStatus:
          type: string
          description: Текущий статус отклика студента
          example: INTERVIEW
        closed:
          type: boolean
          description: Закрыт ли диалог для отправки сообщений
          example: false
    MessageResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор сообщения
          example: 15
        senderId:
          type: integer
          format: int64
          description: Идентификатор отправителя
          example: 7
        senderName:
          type: string
          description: Имя отправителя
          example: Арина
        text:
          type: string
          description: Текст сообщения
          example: Спасибо, мне удобно завтра после 15:00.
        createdAt:
          type: string
          description: Дата и время отправки в пользовательском формате
          example: 18.05.2026 15:10
        mine:
          type: boolean
          description: Отправлено ли сообщение текущим пользователем
          example: true
  responses:
...
    MessageValidationError:
      description: Ошибка отправки сообщения
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            blank:
              summary: Пустое сообщение
              value:
                message: Проверьте корректность заполнения формы
                errorCode: VALIDATION_ERROR
                fields:
                  text: Текст сообщения не может быть пустым
    UnauthorizedError:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              summary: Требуется вход
              value:
                message: Необходимо войти в аккаунт
                errorCode: UNAUTHORIZED
                fields: {}
    NotFoundError:
      description: Сущность не найдена
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              summary: Чат не найден
              value:
                message: Чат не найден
                errorCode: NOT_FOUND
                fields: {}
...
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal:
              summary: Непредвиденная ошибка
              value:
                message: Что-то пошло не так. Попробуйте позже
                errorCode: INTERNAL_ERROR
                fields: {}